feat: refactor SideNav splitting up separate components
